restructuring, moving components into separate folders
