+++ source/components/Developer/ShowCharacterData.ts
@@ -6,8 +6,8 @@ import ICharacter from "@utilities/ICharacter.js";
 import Config from "@utilities/Config.js";
 import { getDataFromFile, saveDataToFile } from "@utilities/StorageService.js";
 import { inputValidators } from "@utilities/MenuService.js";
-import { themedSelectInRoom } from "./ThemedSelectInRoom.js";
-import { themedInput } from "./ThemedInput.js";
+import { themedSelectInRoom } from "../General/ThemedSelectInRoom.js";
+import { themedInput } from "../General/ThemedInput.js";
 
 const getCharacterOptions = (character: ICharacter) => {
   // Calculate inventory sum
