+++ source/components/General/InspectCharacter.ts
