+++ source/components/General/InspectInventory.ts
