+++ source/utilities/GameService.ts
@@ -114,7 +114,7 @@ async function handleItemUsage(characterData: any): Promise<void> {
   ];
 
   const { themedSelectInRoom } = await import(
-    "@components/ThemedSelectInRoom.js"
+    "@components/General/ThemedSelectInRoom.js"
   );
 
   // Show main inventory menu
@@ -904,7 +904,7 @@ export async function startCampaign(): Promise<void> {
       return;
     }
     const { themedSelectInRoom } = await import(
-      "@components/ThemedSelectInRoom.js"
+      "@components/General/ThemedSelectInRoom.js"
     );
     const paceChoice = await themedSelectInRoom({
       message: chalk.hex(getTheme().accentColor)("Choose your story pace:"),
@@ -1010,7 +1010,8 @@ export async function startCampaign(): Promise<void> {
  * Difficulty should be the level of the player, or relative to it
  */
 export function getRandomEnemy(difficulty: number): IEnemy {
-  const maxhp = difficulty > 10 ? Math.floor(difficulty * getRandomNumber(0.75, 1.25)) : 10;
+  const maxhp =
+    difficulty > 10 ? Math.floor(difficulty * getRandomNumber(0.75, 1.25)) : 10;
   const hp = getRandomNumber(maxhp * 0.5, maxhp);
   const name = getEnemyName(difficulty);
