+++ source/utilities/StorageService.ts
@@ -4,7 +4,7 @@ import { log } from "@utilities/LogService.js";
 import fs from "fs-extra";
 import { inputValidators } from "@utilities/MenuService.js";
 import { getTerm } from "@utilities/LanguageService.js";
-import { themedInput } from "@components/ThemedInput.js";
+import { themedInput } from "@components/General/ThemedInput.js";
 
 const files = {
   character: Config.CHARACTER_FILE,
